storage: add 2-phase deletes for finalizers
This is VERY loosely adapted from the K8s flows (sources linked
in code comments).

The general flows:
 * Delete request received for object with no finalizers
    1. Delete immediately; nothing more to do

 * Delete request received for object with finalizers
    1. Set `DeletionTimestamp` + `DeletionGracePeriodSeconds`
        * `DeletionTimestamp` is `now()` (assuming it's nil or in the
          future)
        * `DeletionGracePeriodSeconds` is always 0
    2. Send update notification to trigger reconcilation
    3. Controller reconcilers execute
        * Because `DeletionTimestamp` is set, finalizer(s) run and
          removed from object; object is updated
    4. Update request received (by apiserver)
        * If no more finalizers (on updated obj) and deletion timestamp
          is set, delete object and send delete notification
        * If still more finalizers, follow normal update process

This allows finalizers to drain before the object is fully deleted.
The weirdest part is that the final delete happens as part of an update,
but that mirrors the K8s logic, so I'm not sure we can do better here.

Local tests with `kubectl` show that this flow does work as expected,
with the CLI blocking until the finalizers have run and object is actually
deleted.