storage: fix optimistic concurrency check for subresources (#52)
Subresources are not their own object and share version with their
parent. Subresource updates (e.g. `PUT /.../status`) are really
more as a form of RBAC to avoid unintentionally changing the spec.

As a result, the `builderrest.StatusSubResourceStrategy` works by
copying the status from the updated object to the current storage
object so that inherits everything from the parent object EXCEPT
status, which includes resource version. As a result, we need to
do the optimistic concurrency check (compare versions) BEFORE this
happens.