storage: prevent stale writes in MemoryFS (#53)
Hold object versions in memory and only accept writes if the
version passed matches the object storage version.

Currently, this is needed in addition to the already existing
optimistic concurrency logic due to how we calculate resource versions
currently (global counter, not per-object). We can refactor in the
future to handle this closer to the upstream implementation that
relies on etcd where we capture the storage version up-front (to
pass down to storage layer, exactly like we're doing now), but then
use the various hooks to remove the version before storage and then
re-hydrate it on read.

For now, this is sufficient to prevent bugs at the expense of a bit
of duplicated code.

Currently, this only applies to `MemoryFS` and not `RealFS`, but
in practice, that's fine for right now because of the usage patterns
of those objects currently, which do not have many concurrent writers.