storage: provide guaranteed update for optimistic concurrency retry
There's two classes of optimistic concurrency failures:
 1. Object update is outdated by time it's received by server
 2. Object is written to between update start and FS write

The first case should result in an immediate error propagated to
the user - there's no automatic resolution possible on the server.
(This is largely the optimistic concurrency check we've had thus
far.)

The second one is trickier: the first step at addressing this was
done in #53 (d5d681f238e029bfab17c270e3189c09d2d26497), which at
prevents two writes for the same version from succeeding. However,
this means that patch operations can fail due to optimistic
concurrency, which is explicitly against the apiserver contract.

As a result, this adds a "guaranteed update" method (name taken
from the etcd code that does logically the same thing). This will
retry on any FS/storage-level optimistic concurrency issues, by
replaying the whole update attempt from object fetch.

For an object replace, this means if it's valid at operation start
but another operation beats it before it's persisted, it'll retry,
fetch the new version of the object (or `nil` if object was deleted!)
and then trigger the logic to detect (1), as there will now be a
version mismatch _within the update logic_.

For an object patch, the patch will get replayed on the new object
version and persistence re-attempted. This will continue indefinitely
until a write actually succeeds or the context is canceled.

Additionally, to support this, object versioning has been overhauled
so that it's tracked per-object. For `MemoryFS`, it behaves like etcd
storage, where the versions are removed pre-serialization and restored
on read. Internally, the map now holds a tuple of the version and the
object data (eliminating the separate versions map to simplify book
keeping). This also enables an optimization to avoid incrementing the
version if the object hasn't changed at all. This is useful to avoid
unnecessary optimistic concurrency failures (client A submits no-op
update from v1, client B submits real update from v1, previously if
client B was processed after A, it would fail, now it'll still succeed
since A's no-op update will _really_ be a no-op.) This also matches
K8s apiserver/etcd behavior.

TL;DR Optimistic concurrency is respected throughout the entire
 update lifecycle AND patch works as expected!